Unify app options menus
