+++ packages/toolpad-app/src/server/data.ts
@@ -529,14 +529,19 @@ export async function loadRenderTree(appId: string, version: VersionOrPreview = 
   return appDom.createRenderTree(await loadDom(appId, version));
 }
 
-export async function duplicateApp(id: string, name: string): Promise<AppMeta> {
+export async function duplicateApp(
+  id: string,
+  name: string,
+  suppliedName?: boolean,
+): Promise<AppMeta> {
   const dom = await loadPreviewDom(id);
   const duplicateCount = await prismaClient.app.count({
     where: { name: { startsWith: `${name} (copy`, endsWith: ')' } },
   });
+
   const duplicateName =
     duplicateCount === 0 ? `${name} (copy)` : `${name} (copy ${duplicateCount + 1})`;
-  const newApp = await createApp(duplicateName, {
+  const newApp = await createApp(suppliedName ? name : duplicateName, {
     from: {
       kind: 'dom',
       dom,
