+++ packages/toolpad-app/src/toolpad/AppOptions/AppNameEditable.tsx
@@ -1,0 +1,69 @@
+import * as React from 'react';
+import Skeleton from '@mui/material/Skeleton';
+import type { AppMeta } from '../../server/data';
+import EditableText from '../../components/EditableText';
+import { errorFrom } from '../../utils/errors';
+import client from '../../api';
+
+interface AppNameEditableProps {
+  app?: AppMeta;
+  editing?: boolean;
+  setEditing: (editing: boolean) => void;
+  loading?: boolean;
+}
+
+const AppNameEditable = ({ app, editing, setEditing, loading }: AppNameEditableProps) => {
+  const [appRenameError, setAppRenameError] = React.useState<Error | null>(null);
+  const appNameInput = React.useRef<HTMLInputElement | null>(null);
+  const [appName, setAppName] = React.useState<string>(app?.name || '');
+
+  const handleAppNameChange = React.useCallback(
+    (newValue: string) => {
+      setAppRenameError(null);
+      setAppName(newValue);
+    },
+    [setAppName],
+  );
+
+  const handleAppRenameClose = React.useCallback(() => {
+    setEditing(false);
+    setAppRenameError(null);
+  }, [setEditing]);
+
+  const handleAppRenameSave = React.useCallback(
+    async (name: string) => {
+      if (app?.id) {
+        try {
+          await client.mutation.updateApp(app.id, { name });
+          await client.invalidateQueries('getApps');
+        } catch (rawError) {
+          setAppRenameError(errorFrom(rawError));
+          setEditing(true);
+        }
+      }
+    },
+    [app?.id, setEditing],
+  );
+
+  return loading ? (
+    <Skeleton />
+  ) : (
+    <EditableText
+      defaultValue={app?.name}
+      editable={editing}
+      helperText={appRenameError ? `An app named "${appName}" already exists` : null}
+      error={!!appRenameError}
+      onChange={handleAppNameChange}
+      onClose={handleAppRenameClose}
+      onSave={handleAppRenameSave}
+      ref={appNameInput}
+      sx={{
+        width: '100%',
+      }}
+      value={appName}
+      variant="subtitle1"
+    />
+  );
+};
+
+export default AppNameEditable;
